+++ packages/eslint-plugin-pf-codemods/test/rules/v5/react-icons-warn-remove-props.js
@@ -1,5 +1,5 @@
 const ruleTester = require("../../ruletester");
-const rule = require("../../../lib/rules/v5/react-icons-remove-props");
+const rule = require("../../../lib/rules/v5/react-icons-warn-remove-props");
 
 ruleTester.run("react-icons-remove-props", rule, {
   valid: [
@@ -14,18 +14,9 @@ ruleTester.run("react-icons-remove-props", rule, {
   invalid: [
     {
       code: `import { FrogIcon } from '@patternfly/react-icons'; <FrogIcon size color noVerticalAlign />`,
-      output: `import { FrogIcon } from '@patternfly/react-icons'; <FrogIcon    />`,
       errors: [
         {
-          message: `size prop for FrogIcon has been removed`,
-          type: "JSXOpeningElement",
-        },
-        {
-          message: `color prop for FrogIcon has been removed`,
-          type: "JSXOpeningElement",
-        },
-        {
-          message: `noVerticalAlign prop for FrogIcon has been removed`,
+          message: `The size, color, and noVerticalAlign properties have been removed from react-icons. Wrap your react-icon with the <Icon> component to customize its size and color.`,
           type: "JSXOpeningElement",
         },
       ],
