+++ packages/pf-codemods/README.md
@@ -1799,23 +1799,11 @@ The `react-dropzone` dependency used with FileUpload, MultipleFileUpload, and Co
 
 This rule will raise a warning when any of these three components are imported, but will not make any code changes.
 
-### react-icons-remove-props [(#5275)](// https://github.com/patternfly/patternfly-react/pull/5275)
+### react-icons-warn-remove-props [(#5275)](// https://github.com/patternfly/patternfly-react/pull/5275)
 
-We've removed the `size`, `color`, and `noVerticalAlign` properties from icons which are imported from @patternfly/react-icons.
+We've removed the `size`, `color`, and `noVerticalAlign` properties from icons which are imported from @patternfly/react-icons. Wrap your react-icon with the <Icon> component to customize its size and color.
 
-#### Examples
-
-In:
-
-```jsx
-<FrogIcon size="sm" color="green" noVerticalAlign />
-```
-
-Out:
-
-```jsx
-<FrogIcon     />
-```
+This rule will raise a warning when any of these three props are used, but will not make any code changes.
 
 ### remove-removeFindDomNode [(#8371)](https://github.com/patternfly/patternfly-react/pull/8371) [(#8316)](https://github.com/patternfly/patternfly-react/pull/8316)
